Refactored entire project.
