+++ index.css
@@ -13,6 +13,11 @@ b{
   font-weight: normal;
 }
 
+.hidden
+{
+  display: none;
+}
+
 :root {
   --off-white: rgb(221, 221, 221);
   --light-off-white: rgb(241, 241, 241);
