+++ index.js
@@ -1,516 +1,46 @@
-import { showNotification } from "./js/utils/userActions/showNotification.js";
-import db from "./js/dexie/dexie.js";
-import fetchSuttaData from './js/database/fetchSuttaData.js';
-import getDocumentAreas from './js/utils/getDocumentAreas.js';
-import { getSuttasByIds } from './js/utils/getSuttasByIds.js';
-import { goBack } from './js/utils/navigation/goBack.js';
-import { showForeword } from './js/utils/userActions/showForeword.js';
-import { scrollToHash } from './js/utils/navigation/scrollToHash.js';
+import { loadWhatsNewArea } from "./js/utils/contentSections/loadWhatsNewArea.js";
+import { displaySuttas } from "./js/utils/contentSections/displaySuttas.js";
+import { activateEventListeners } from "./js/utils/loadContent/activateEventListeners.js";
+import { fetchAvailableSuttas } from "./js/utils/loadContent/fetchAvailableSuttas.js";
+import initializeSideBySide from "./js/utils/loadContent/initializeSidebyside.js";
+import { toggleTheme } from "./js/utils/misc/toggleTheme.js";
+import { buildSutta } from "./js/utils/loadContent/buildSutta.js";
+import { handleFetchSuttaTranslations } from "./js/utils/async/handleFetchSuttaTranslations.js";
+import updateSuttaDatabase from './js/database/updateSuttaDatabase.js';
 
 
-const {suttaArea, whatsNewArea, homeButton, themeButton, bodyTag, previous, next} 
-  = getDocumentAreas();
 
-fetchSuttaData();
+// Wait for DOM to be fully loaded
+document.addEventListener('DOMContentLoaded', async function() {
+  try {
+      // Fetch data and initialize app
+      const initialThemeSetting = localStorage.theme === "dark";
 
+      const availableSuttasJson = await fetchAvailableSuttas();
+      activateEventListeners(availableSuttasJson);
 
-function removeDiacritics(text) {
-  if (!text) return "";
-  return text.normalize("NFD").replace(/[\u0300-\u036f]/g, "");
-}
-
-function loadWhatsNewArea() {
-  function daysAgo(dateString) {
-    const dateAdded = new Date(dateString);
-    const currentDate = new Date();
-    const timeDiff = Math.abs(currentDate - dateAdded);
-    const daysDiff = Math.floor(timeDiff / (1000 * 60 * 60 * 24)); // Convert time difference to days
-    return daysDiff === 0 ? 'Today' : `${daysDiff} days ago`;
-  }
-
-  // Sort the suttas by date_added in descending order
-  const sortedSuttas = Object.values(availableSuttasJson).sort((a, b) => new Date(b.date_added) - new Date(a.date_added));
-  const recentSuttas = sortedSuttas.slice(0, 5);
-
-  if (whatsNewArea) {
-    whatsNewArea.innerHTML = `<h2>What's New</h2>` + 
-    `<div class="whats-new-container">
-        ${recentSuttas.map(sutta => {
-          const id = sutta.id.replace(/\s+/g, '');
-          const title = sutta.title;
-          const daysAgoAdded = daysAgo(sutta.date_added); // Calculate how many days ago it was added
-          const link = `<a href="/?q=${id.toLowerCase()}">${title}</a>`;
-          return `
-            <div class="sutta-box">
-              <h3 class="sutta-card-title">${link}</h3>
-              <div class="sutta-pali-title"><em>${id}: ${sutta.pali_title}</em></div>
-              <div class="sutta-date-added"><small>Added: ${daysAgoAdded}</small></div>
-            </div>
-          `;
-        }).join('')}
-      </div>
-    `;
-  }
-}
-
-function displaySuttas(suttas, isSearch = false) {
-  // foreword button
-  const forewordViewed = localStorage.getItem('forewordViewed', false);
-  const forewordButton = document.getElementById('foreword-button');
-
-  if (forewordViewed == 'true' && forewordButton) {
-    forewordButton.style.display = 'none';
-  }
-
-  // display all suttas
-  const books = {
-    "dn": "Dīgha Nikāya",
-    "mn": "Majjhima Nikāya",
-    "sn": "Saṃyutta Nikāya",
-    "an": "Aṅguttara Nikāya",
-    "kn": "Khuddaka Nikāya"
-  };
-  let currentGroup = -1;
-
-  if (isSearch) {
-    suttaArea.innerHTML += "<h2>Search Results:</h2>";
-    whatsNewArea.style.display = "none";
-  } else {
-    whatsNewArea.style.display = "block";
-  }
-
-  suttaArea.innerHTML += `<ul style="margin-top: 20px;">${Object.entries(suttas).map(([sutta_id, sutta_details]) => {
-    const id = sutta_details['id'].replace(/\s+/g, '');;
-    const title = sutta_details['title']
-    const heading = sutta_details['heading'] || ""
-    const link = `<a href="/?q=${id.toLowerCase()}">${id}: ${title}`;
-    const em = heading ? `<span style="color: #7f6e0a;"><em>${heading}</em></span>` : '';
-    const nikaya = sutta_id.slice(0, 2).toLowerCase();
-
-    // Check if the current sutta belongs to a new group
-    const key = Object.keys(books)[currentGroup];
-
-    if (!isSearch && nikaya !== key && currentGroup < 4) {
-      // If it's a new group, display the subheading
-      currentGroup += 1;
-      const key = Object.keys(books)[currentGroup];
-
-      return `<h2>${books[key]}</h2><li>${link}${em ? ` (${em})` : ''}</a></li>`;
-    } else {
-      return `<li>${link}${em ? ` (${em})` : ''}</a></li>`;
-    }
-  }).join('')}</ul>`;
-}
-
-function initializePaliToggle() {
-  const hideButton = document.getElementById("hide-pali");
-  if (localStorage.paliToggle !== "show") {
-    localStorage.paliToggle = "hide";
-    suttaArea.classList.add("hide-pali");
-  }
-  hideButton.addEventListener("click", () => {
-    const englishElements = Array.from(suttaArea.querySelectorAll(".eng-lang"));
-    const firstVisibleEnglishElement = englishElements.find(el => {
-      const rect = el.getBoundingClientRect();
-      return rect.bottom >= 0 && rect.top <= window.innerHeight;
-    });
-    const togglePali = () => {
-      if (localStorage.paliToggle === "show") {
-        suttaArea.classList.add("hide-pali");
-        localStorage.paliToggle = "hide";
-        document.body.classList.remove("side-by-side");
-        localStorage.sideBySide = "false";
+      // Initialize based on URL or default content
+      if (document.location.search) {
+          const slug = document.location.search.replace("?q=", "").replace(/\s/g, "").replace(/%20/g, "");
+          buildSutta(slug, availableSuttasJson);
       } else {
-        suttaArea.classList.remove("hide-pali");
-        localStorage.paliToggle = "show";
+          displaySuttas(availableSuttasJson);
+          loadWhatsNewArea(availableSuttasJson);
       }
-    };
-    if (firstVisibleEnglishElement) {
-      const prevOffset = firstVisibleEnglishElement.getBoundingClientRect().top;
-      togglePali();
-      const newOffset = firstVisibleEnglishElement.getBoundingClientRect().top;
-      window.scrollBy(0, newOffset - prevOffset);
-    } else {
-      togglePali();
-    }});
-}
 
-// Event listeners
+      // Additional setup steps
+      await handleFetchSuttaTranslations();
+      updateSuttaDatabase();
+      initializeSideBySide();
+      toggleTheme(initialThemeSetting);
 
-// Toggle side-by-side mode
-document.onkeyup = function (e) {
-  const paliHidden = document.getElementById("sutta").classList.contains("hide-pali");
-  if (!paliHidden && e.target.id != "search-bar" && e.key == "s") {
-    if (localStorage.sideBySide === "true") {
-      bodyTag.classList.remove("side-by-side");
-      localStorage.sideBySide = "false";
-    } else {
-      bodyTag.classList.add("side-by-side");
-      localStorage.sideBySide = "true";
-    }
-  }
-};
+      // Remove the hidden class once everything is done
+      document.getElementById('appbody').classList.remove('hidden');
 
-homeButton.addEventListener("click", () => {
-  window.location.href = '/';
-});
+  } catch (error) {
+      console.error('[ERROR] Something went wrong:', error);
 
-var converter = new showdown.Converter();
-
-const response = await fetch('available_suttas.json');
-const availableSuttas = await response.json();
-const availableSuttasJson = availableSuttas['available_suttas'];
-
-
-// initialize
-if (localStorage.sideBySide) {
-  if (localStorage.sideBySide == "true") {
-    bodyTag.classList.add("side-by-side");
-  }
-} else {
-  bodyTag.classList.remove("side-by-side");
-}
-// Function to toggle theme
-function toggleTheme(useDarkTheme) {
-  bodyTag.classList.remove(useDarkTheme ? "light" : "dark");
-  bodyTag.classList.add(useDarkTheme ? "dark" : "light");
-  localStorage.theme = useDarkTheme ? "dark" : "light";
-}
-
-// Set initial theme
-const initialUseDarkTheme = localStorage.theme === "dark";
-toggleTheme(initialUseDarkTheme);
-
-// Event listener for theme toggle
-themeButton.addEventListener("click", () => {
-  const currentThemeIsDark = localStorage.theme === "dark";
-  toggleTheme(!currentThemeIsDark);
-});
-
-// input in search bar
-const searchBar = document.getElementById("search-bar");
-searchBar.focus();
-
-searchBar.addEventListener("input", async (e) => {
-  const searchQuery = e.target.value.trim().toLowerCase();
-  if (!searchQuery) {
-    suttaArea.innerHTML = "";
-    displaySuttas(availableSuttasJson);
-    return;
-  }
-
-  const collection = db.suttas.filter((sutta) => {
-    const suttaContent = JSON.stringify(sutta.value).toLowerCase();
-    if (suttaContent.includes(searchQuery)) return true;
-
-    const suttaContentWithoutDiacritics = removeDiacritics(suttaContent);
-    return suttaContentWithoutDiacritics.includes(searchQuery);
-  });
-  const searchResults = await collection.toArray();
-  const suttaIds = searchResults.map((result) => result.id)
-  const suttas = getSuttasByIds(suttaIds, availableSuttasJson);
-  if (suttaIds.length > 0) {
-    suttaArea.innerHTML = "";
-    displaySuttas(suttas, true);
-  }
-  else {
-    suttaArea.innerHTML = "<h2 class=\"no-results\">No results found</h2>";
+      // If something fails, still reveal the default content
+      document.getElementById('body').classList.remove('hidden');
   }
 });
-
-document.getElementById("form").addEventListener("submit", e => {
-  e.preventDefault();
-  const searchValue = searchBar.value.trim().replace(/\s/g, "");
-  if (searchValue) {
-    buildSutta(searchValue);
-    history.pushState({ page: searchValue }, "", `?q=${searchValue}`);
-  }
-});
-
-function buildSutta(slug) {
-  slug = slug.toLowerCase();
-  let sutta_details = availableSuttasJson[slug]
-  let translator = "Bhikkhu Anīgha";
-  let html = `<div class="button-area"><button id="hide-pali" class="hide-button">Toggle Pali</button></div>`;
-  // TODO if file names are consistent we can get it from the availablesuttasjson
-  let sutta_title = sutta_details['title'];
-
-  const rootResponse = fetch(sutta_details['root_path']).then(response => response.json());
-  const translationResponse = fetch(sutta_details['translation_path']).then(response => response.json());
-  const htmlResponse = fetch(sutta_details['html_path']).then(response => response.json());
-  const commentResponse = fetch(sutta_details['comment_path'])
-    .then(response => {
-      if (!response.ok) throw new Error(`Comment file not found for ${slug}`);
-      return response.json();
-    })
-    .catch(error => {
-      console.warn(error.message);
-      return {}; // Return an empty object if the comment file is not found
-    });
-
-  const authors = fetch(`authors.json`).then(response => response.json());
-
-  // Get root, translation and html jsons from folder
-  Promise.all([htmlResponse, rootResponse, translationResponse, commentResponse, authors])
-    .then(responses => {
-      const [html_text, root_text, translation_text, comment_text, authors_text] = responses;
-      const keys_order = Object.keys(html_text);
-      let commentCount = 1;
-      let commentsHtml = '';
-      keys_order.forEach((segment) => {
-        if (translation_text[segment] === undefined) {
-          translation_text[segment] = "";
-        }
-        let [openHtml, closeHtml] = html_text[segment].split(/{}/);
-        // openHtml = openHtml.replace(/^<span class='verse-line'>/, "<br><span class='verse-line'>");
-
-        if (window.addBreaks === true) {
-          openHtml = openHtml.replace(/^<span class='verse-line'>/, "<br><span class='verse-line'>");
-        }
-
-        html +=
-          `${openHtml}<span class="segment" id="${segment}">` +
-          `<span class="pli-lang" lang="pi">${root_text[segment] || ''}</span>` +
-          `<span class="eng-lang" lang="en">${translation_text[segment]}` +
-          `${comment_text[segment] ? `<a href="#comment${commentCount}" class="comment">[${commentCount}]</a>` : ''}` +
-          `</span></span>${closeHtml}\n\n`;
-        
-          if (comment_text[segment]) {
-            if(commentCount == 1){
-              commentsHtml += '<h3>Comments</h3>';
-            }
-            // Inside the comment HTML
-            commentsHtml += `
-            <p id="comment${commentCount}"><span>
-              ${commentCount}: ${converter.makeHtml(comment_text[segment])
-                .replace(/^<p>(.*)<\/p>$/, '$1')}
-              <a href="#${segment}~no-highlight" style="cursor: pointer; font-size: 14px;">&larr;</a>
-            </span></p>
-            `;
-  
-            commentCount++;
-          }
-      });
-
-      if (authors_text[slug]) translator = authors_text[slug];
-      const translatorByline = `<div class="byline"><p>Translated by ${translator}</p></div>`;
-      // render comments
-      
-      suttaArea.innerHTML = `<p class="sc-link"></p>` + html + translatorByline + commentsHtml;
-
-      let acronym = sutta_details['id'];
-
-      document.title = `${acronym} ` + sutta_title;
-
-      initializePaliToggle();
-      addNavbar();
-
-      // remove download and info button
-      const cacheButton = document.getElementById('cacheButton');
-      const infoButton = document.getElementById('infoButton');
-      const epubInfoButton = document.getElementById('epubInfoButton');
-      if (cacheButton) cacheButton.style.display = 'none';
-      if (infoButton) infoButton.style.display = 'none';
-      if (epubInfoButton) epubInfoButton.style.display = 'none';
-
-      // scroll to the quote in the url if present
-      scrollToHash();
-    })
-    .catch(error => {
-      suttaArea.innerHTML = `<p>Sorry, "${decodeURIComponent(slug)}" is not a valid sutta citation.
-
-    <br><br>Note: Make sure the citation code is correct. Otherwise try finding the sutta from the home page.<br>`;
-    });
-}
-
-function addNavbar() {
-  // Add the navbar to the page
-  const navbar = document.createElement('div');
-  navbar.id = 'suttanav'; // Added ID
-  navbar.innerHTML = document.title;
-  document.body.appendChild(navbar);
-
-  let lastScrollTop = 0; // variable to store the last scroll position
-  const scrollThreshold = 10;
-  let isScrolling = false;
-  let scrollTimeout;
-
-  window.addEventListener('scroll', () => {
-    if (!isScrolling) {
-      isScrolling = true;
-      requestAnimationFrame(() => {
-        let currentScrollTop = window.scrollY || document.documentElement.scrollTop;
-
-        // Detect sudden jump
-        if (Math.abs(currentScrollTop - lastScrollTop) > 100) {
-          // If the jump is large, do not show the navbar
-          navbar.style.top = '-50px';
-        } else if (Math.abs(currentScrollTop - lastScrollTop) > scrollThreshold) {
-          // Only apply the scroll behavior if it's a regular scroll (not a sudden jump)
-          navbar.style.top = currentScrollTop < 170 || currentScrollTop > lastScrollTop ? '-50px' : '0';
-        }
-
-        lastScrollTop = currentScrollTop;
-        isScrolling = false;
-      });
-    }
-
-    // Clear any previous timeout
-    clearTimeout(scrollTimeout);
-
-    // Set a timeout to handle the case when the user stops scrolling
-    scrollTimeout = setTimeout(() => {
-      isScrolling = false;
-    }, 100);
-  });
-}
-
-
-// initialize the whole app
-if (document.location.search) {
-  buildSutta(document.location.search.replace("?q=", "").replace(/\s/g, "").replace(/%20/g, ""));
-} else {
-  displaySuttas(availableSuttasJson);
-  loadWhatsNewArea();
-}
-
-
-document.addEventListener('click', function (event) {
-  // Check if the clicked element is the foreword button
-  if (event.target && event.target.id === 'foreword-button') {
-    showForeword(); // Call the function to show the foreword
-    displaySuttas(availableSuttasJson);
-  }
-  // Add a click event listener to the span element
-  if (event.target && event.target.id === 'back-arrow') {
-    goBack();
-  }
-});
-
-const refreshButton = document.getElementById("hardRefresh");
-refreshButton.addEventListener("click", function () {
-  db.delete().then(() => {
-    console.log("[SUCCESS] Database deleted");
-    localStorage.clear();
-    window.location.href = "/";
-  }).catch((error) => {
-    console.error("[ERROR] Could not delete database:", error);
-  })
-});
-
-
-const errorButton = document.getElementById('reportButton');
-errorButton.addEventListener('click', () => {
-  window.open('https://docs.google.com/forms/d/1Ng8Csf9xYJ7UaYUyl3sGEyZ3aa2FJE_0GRS6zI6oIBM/edit', '_blank');
-});
-
-window.addEventListener('hashchange', function () {
-  const hash = window.location.hash;
-
-  // Check if the hash starts with "comment"
-  if (hash.startsWith('#comment')) {
-    return; // Let the browser handle the default behavior
-  }
-
-  scrollToHash();
-});
-
-document.getElementById('cacheButton').addEventListener('click', () => {
-  // Check if service worker is supported by the browser
-  if ('serviceWorker' in navigator) {
-    // Send message to service worker to trigger caching
-    try {
-      showNotification("Downloading...")
-      navigator.serviceWorker.controller.postMessage({ action: 'cacheResources' });
-    } catch (error) {
-      console.log(error);
-      // TODO maybe a red colour box here?
-      showNotification("An error occurred while attempting to download. Please refresh the page, wait a few seconds, and retry");
-    }
-  }
-});
-
-infoButton.addEventListener("click", function (event) {
-  event.stopPropagation(); // Prevent click from immediately propagating to document
-  let notificationBox = document.querySelector('.info-notification-box')
-  if (!notificationBox) {
-    notificationBox = document.createElement('div');
-    notificationBox.classList.add('info-notification-box');
-    document.body.appendChild(notificationBox);
-  }
-
-  if (notificationBox.style.display == 'block') {
-    notificationBox.style.display = 'none';
-  } else {
-    notificationBox.textContent = "The ‘Use Offline’ button makes the site available offline on the current web browser at the same URL (suttas.hillsidehermitage.org).\n\nThe site can also be installed as an application on mobile phones, by tapping ‘Install’ at the menu on the top right corner. Note that hitting the ‘Use Offline’ button is still necessary to make it available offline through the app.\n\nIf downloading again (e.g., when new Suttas become available), make sure to first clear the site's data on your browser/app and reload the page.";
-    notificationBox.style.display = 'block';
-  }
-});
-// Add event listener to document to hide notificationBox when clicking outside
-document.addEventListener("click", function (event) {
-  let notificationBox = document.querySelector('.info-notification-box');
-  if (notificationBox && notificationBox.style.display == 'block') {
-    // Check if the click is outside the notificationBox and not on the infoButton
-    if (!notificationBox.contains(event.target) && event.target !== infoButton) {
-      notificationBox.style.display = 'none';
-    }
-  }
-});
-
-let lastModifiedDate;
-epubInfoButton.addEventListener("click", function (event) {
-  event.stopPropagation(); // Prevent click from immediately propagating to document
-  let notificationBox = document.querySelector('.info-notification-box')
-  if (!notificationBox) {
-    notificationBox = document.createElement('div');
-    notificationBox.classList.add('info-notification-box');
-    document.body.appendChild(notificationBox);
-  }
-
-  if (notificationBox.style.display == 'block') {
-    notificationBox.style.display = 'none';
-  } else {
-    notificationBox.textContent = "The ‘Get Ebook’ button lets you download the translations and comments in an Ebook with the ‘.epub’ format." + (lastModifiedDate != null ? "\n\nLast updated: " + lastModifiedDate : "");
-    notificationBox.style.display = 'block';
-  }
-});
-
-document.getElementById('downloadEpubButton').addEventListener('click', function() {
-    const link = document.createElement('a');
-    link.href = '/suttas_epub/Sutta_Translations.epub';  // Path to EPUB file
-    link.download = 'Sutta_Translations.epub';    // Downloaded file name
-    link.click();
-});
-
-navigator.serviceWorker.addEventListener('message', event => {
-  if (event.data && event.data.action === 'cachingSuccess') {
-    showNotification("Download successful - site available offline.")
-  }
-  if (event.data && event.data.action === 'cachingError') {
-    // TODO again maybe a different colour box
-    showNotification("Caching error. Please clear site data, refresh the page, and try again.");
-  }
-});
-
-fetch('/suttas_epub/Sutta_Translations.epub')
-  .then(response => {
-    if (response.ok) {
-      const lastModified = response.headers.get('Last-Modified');
-      if (lastModified) {
-        const date = new Date(lastModified);
-        // Format date
-        lastModifiedDate = date.toLocaleDateString('en-EN', {
-          year: 'numeric',
-          month: 'long',
-          day: 'numeric'
-        });
-      } else {
-        console.error("Header 'Last-Modified' is missing or invalid.");
-      }
-    } else {
-      console.error('Error when trying to get file :', response.statusText);
-    }
-  })
-  .catch(error => {
-    console.error('Error :', error);
-  });
