+++ js/utils/eventListeners/activateHomeButton.js
@@ -1,0 +1,11 @@
+import getDocumentAreas from "../getDocumentAreas.js";
+
+export function activateHomeButton()
+{
+    const {homeButton} = getDocumentAreas();
+
+    homeButton.addEventListener("click", () => 
+    {
+        window.location.href = '/';
+    });
+}
