+++ js/utils/eventListeners/activateMessageListener.js
@@ -1,0 +1,15 @@
+
+
+export default function activateMessageListener()
+{
+    const SUCCESS_MSG = "Download successful - site available offline.";
+    const FAIL_MSG = "Caching error. Please clear site data, refresh the page, and try again."
+
+
+    navigator.serviceWorker.addEventListener('message', event => 
+    {
+        if(!event.data) return;
+        if (event.data.action === 'cachingSuccess') showNotification(SUCCESS_MSG)
+        if (event.data.action === 'cachingError') showNotification(FAIL_MSG);
+    });
+}
