+++ js/utils/eventListeners/activateRefershButton.js
@@ -1,0 +1,17 @@
+
+
+export default function activateRefreshButton()
+{
+    const refreshButton = document.getElementById("hardRefresh");
+    refreshButton.addEventListener("click", function () 
+    {
+        db.delete().then(() => {
+            console.log("[SUCCESS] Database deleted");
+            localStorage.clear();
+            window.location.href = "/";
+        }).catch((error) => 
+        {
+            console.error("[ERROR] Could not delete database:", error);
+        })
+    });
+}
