+++ js/utils/eventListeners/activateThemeButton.js
@@ -1,0 +1,13 @@
+import getDocumentAreas from "../getDocumentAreas.js";
+
+export function activateThemeButton()
+{
+    const {themeButton} = getDocumentAreas();
+
+    themeButton.addEventListener("click", () => 
+    {
+        const currentThemeIsDark = localStorage.theme === "dark";
+        toggleTheme(!currentThemeIsDark);
+    });
+    
+}
