+++ js/utils/getDocumentAreas.js
@@ -9,6 +9,20 @@ export default function getDocumentAreas()
     const bodyTag = document.querySelector("body");
     const previous = document.getElementById("previous");
     const next = document.getElementById("next");
+    const searchBar = document.getElementById("search-bar");
+    const form = document.getElementById('form');
+    const infoButton = document.getElementById('infoButton');
 
-    return {suttaArea, whatsNewArea, homeButton, themeButton, bodyTag, previous, next};
+    return {
+        suttaArea, 
+        whatsNewArea, 
+        homeButton, 
+        themeButton, 
+        bodyTag, 
+        previous, 
+        next, 
+        searchBar,
+        form,
+        infoButton,
+    };
 }
