+++ js/utils/misc/removeDiacritics.js
@@ -1,0 +1,4 @@
+export function removeDiacritics(text) {
+    if (!text) return "";
+    return text.normalize("NFD").replace(/[\u0300-\u036f]/g, "");
+}
